split models CLI coverage helpers
