+++ pkg/services/models/transports/cli/root_invoke.go
@@ -11,7 +11,6 @@ import (
 	"strings"
 
 	modelinference "github.com/portpowered/infinite-you/pkg/services/models"
-	"github.com/portpowered/infinite-you/pkg/services/work"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/clidiag"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	contentcontract "github.com/portpowered/infinite-you/pkg/transports/mapping/workcontent"
@@ -997,76 +996,4 @@ func genericInvocationResponseFromInferenceResult(
 		outputs[index] = projected
 	}
 	return factoryapi.GenericModelInvocationResponse{Outputs: outputs}
-}
-
-func genericCLIStringPointer(value string) *string {
-	if strings.TrimSpace(value) == "" {
-		return nil
-	}
-	copy := value
-	return &copy
-}
-
-func catalogPresentationForOperation(catalog modelinference.Detail, operation string) (string, string) {
-	for _, capability := range catalog.Capabilities {
-		for _, catalogOperation := range capability.Operations {
-			if catalogOperation.Name == operation {
-				return capability.Worker, string(capability.ProviderLocality)
-			}
-		}
-	}
-	return "", string(catalog.ProviderLocality)
-}
-
-func resolvedPresentationBindings(
-	catalog modelinference.Detail,
-	operation string,
-	inputText string,
-) []modelinference.ResolvedModelOperationBinding {
-	operationDetail, ok := catalogOperationForName(catalog, operation)
-	if !ok {
-		return []modelinference.ResolvedModelOperationBinding{}
-	}
-	for _, input := range operationDetail.Inputs {
-		slot := strings.TrimSpace(input.Name)
-		if slot == "" {
-			continue
-		}
-		return []modelinference.ResolvedModelOperationBinding{{
-			Slot:   slot,
-			Source: "INPUT",
-			Content: []work.WorkContentPart{{
-				Type: work.WorkContentPartTypeText,
-				Text: inputText,
-			}},
-		}}
-	}
-	return []modelinference.ResolvedModelOperationBinding{}
-}
-
-func catalogOperationForName(catalog modelinference.Detail, operation string) (modelinference.Operation, bool) {
-	for _, catalogOperation := range catalog.Operations {
-		if catalogOperation.Name == operation {
-			return catalogOperation, true
-		}
-	}
-	for _, capability := range catalog.Capabilities {
-		for _, catalogOperation := range capability.Operations {
-			if catalogOperation.Name == operation {
-				return catalogOperation, true
-			}
-		}
-	}
-	return modelinference.Operation{}, false
-}
-
-func catalogCapabilityOperationForName(catalog modelinference.Detail, operation string) (modelinference.Operation, bool) {
-	for _, capability := range catalog.Capabilities {
-		for _, catalogOperation := range capability.Operations {
-			if catalogOperation.Name == operation {
-				return catalogOperation, true
-			}
-		}
-	}
-	return modelinference.Operation{}, false
 }
